requirements.txt
Authlib==1.2.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.3
Flask==2.3.2
Flask-Dance==7.0.0
Flask-FontAwesome==0.1.5
Flask-Login==0.6.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
pycparser==2.21
pyOpenSSL==23.2.0
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
urllib3==2.0.4
URLObject==2.4.3
Werkzeug==2.3.6
